+++ packages/widgetbook/CHANGELOG.md
@@ -3,6 +3,7 @@
 - **BREAKING**: Merge `widgetbook_generator` and `widgetbook_test` packages into `widgetbook` package. ([#1772](https://github.com/widgetbook/widgetbook/pull/1772))
 - **REFACTOR**: Change default font to `Roboto`. ([#1773](https://github.com/widgetbook/widgetbook/pull/1773))
 - **FEAT**: Add `Defaults` class for default `setup` and `builder`. ([#1768](https://github.com/widgetbook/widgetbook/pull/1768))
+- **FEAT**: Create default `Scenario` when none are provided. ([#1775](https://github.com/widgetbook/widgetbook/pull/1775))
 - **FIX**: Show scenarios in the navigation tree. ([#1774](https://github.com/widgetbook/widgetbook/pull/1774))
 
 ## 4.0.0-alpha.4